requirements.txt
python-telegram-bot==13.13
ytmusicapi==0.25.0
pytube==12.1.2
requests~=2.28.2
mutagen~=1.46.0
moviepy~=1.0.3